rag-app/requirements.txt
# requirements.txt
Flask
PyMuPDF
sentence-transformers
faiss-cpu
transformers
torch
requests
beautifulsoup4
markdown2
numpy